@@ -5,7 +5,7 @@ repository = "https://github.com/TheKroni/markov-bot"
 license = "MIT"
 version = "1.0.2"
 authors = ["Kroni"]
-edition = "2018"
+edition = "2021"
 
 [dependencies]
 serenity = { version="0.10.9", default-features = false, features = ["cache", "client", "gateway", "rustls_backend", "model", "unstable_discord_api"] }
